Extract CP775 to its own file
