+++ escpos4k/src/commonMain/kotlin/cz/multiplatform/escpos4k/core/encoding/charset/CP775.kt
@@ -1,0 +1,152 @@
+/*
+ *    Copyright 2024 Ondřej Karmazín
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package cz.multiplatform.escpos4k.core.encoding.charset
+
+/** Page 33 [CP775: Baltic Rim] */
+public data object CP775 : Charset(33, "IBM775") {
+  override val mapping: Map<Char, Int> =
+      mapOf(
+          '\u0106' to 0x80,
+          '\u00FC' to 0x81,
+          '\u00E9' to 0x82,
+          '\u0101' to 0x83,
+          '\u00E4' to 0x84,
+          '\u0123' to 0x85,
+          '\u00E5' to 0x86,
+          '\u0107' to 0x87,
+          '\u0142' to 0x88,
+          '\u0113' to 0x89,
+          '\u0156' to 0x8A,
+          '\u0157' to 0x8B,
+          '\u012B' to 0x8C,
+          '\u0179' to 0x8D,
+          '\u00C4' to 0x8E,
+          '\u00C5' to 0x8F,
+          '\u00C9' to 0x90,
+          '\u00E6' to 0x91,
+          '\u00C6' to 0x92,
+          '\u014D' to 0x93,
+          '\u00F6' to 0x94,
+          '\u0122' to 0x95,
+          '\u00A2' to 0x96,
+          '\u015A' to 0x97,
+          '\u015B' to 0x98,
+          '\u00D6' to 0x99,
+          '\u00DC' to 0x9A,
+          '\u00F8' to 0x9B,
+          '\u00A3' to 0x9C,
+          '\u00D8' to 0x9D,
+          '\u00D7' to 0x9E,
+          '\u00A4' to 0x9F,
+          '\u0100' to 0xA0,
+          '\u012A' to 0xA1,
+          '\u00F3' to 0xA2,
+          '\u017B' to 0xA3,
+          '\u017C' to 0xA4,
+          '\u017A' to 0xA5,
+          '\u201D' to 0xA6,
+          '\u00A6' to 0xA7,
+          '\u00A9' to 0xA8,
+          '\u00AE' to 0xA9,
+          '\u00AC' to 0xAA,
+          '\u00BD' to 0xAB,
+          '\u00BC' to 0xAC,
+          '\u0141' to 0xAD,
+          '\u00AB' to 0xAE,
+          '\u00BB' to 0xAF,
+          '\u2591' to 0xB0,
+          '\u2592' to 0xB1,
+          '\u2593' to 0xB2,
+          '\u2502' to 0xB3,
+          '\u2524' to 0xB4,
+          '\u0104' to 0xB5,
+          '\u010C' to 0xB6,
+          '\u0118' to 0xB7,
+          '\u0116' to 0xB8,
+          '\u2563' to 0xB9,
+          '\u2551' to 0xBA,
+          '\u2557' to 0xBB,
+          '\u255D' to 0xBC,
+          '\u012E' to 0xBD,
+          '\u0160' to 0xBE,
+          '\u2510' to 0xBF,
+          '\u2514' to 0xC0,
+          '\u2534' to 0xC1,
+          '\u252C' to 0xC2,
+          '\u251C' to 0xC3,
+          '\u2500' to 0xC4,
+          '\u253C' to 0xC5,
+          '\u0172' to 0xC6,
+          '\u016A' to 0xC7,
+          '\u255A' to 0xC8,
+          '\u2554' to 0xC9,
+          '\u2569' to 0xCA,
+          '\u2566' to 0xCB,
+          '\u2560' to 0xCC,
+          '\u2550' to 0xCD,
+          '\u256C' to 0xCE,
+          '\u017D' to 0xCF,
+          '\u0105' to 0xD0,
+          '\u010D' to 0xD1,
+          '\u0119' to 0xD2,
+          '\u0117' to 0xD3,
+          '\u012F' to 0xD4,
+          '\u0161' to 0xD5,
+          '\u0173' to 0xD6,
+          '\u016B' to 0xD7,
+          '\u017E' to 0xD8,
+          '\u2518' to 0xD9,
+          '\u250C' to 0xDA,
+          '\u2588' to 0xDB,
+          '\u2584' to 0xDC,
+          '\u258C' to 0xDD,
+          '\u2590' to 0xDE,
+          '\u2580' to 0xDF,
+          '\u00D3' to 0xE0,
+          '\u00DF' to 0xE1,
+          '\u014C' to 0xE2,
+          '\u0143' to 0xE3,
+          '\u00F5' to 0xE4,
+          '\u00D5' to 0xE5,
+          '\u00B5' to 0xE6,
+          '\u0144' to 0xE7,
+          '\u0136' to 0xE8,
+          '\u0137' to 0xE9,
+          '\u013B' to 0xEA,
+          '\u013C' to 0xEB,
+          '\u0146' to 0xEC,
+          '\u0112' to 0xED,
+          '\u0145' to 0xEE,
+          '\u2019' to 0xEF,
+          '\u00AD' to 0xF0,
+          '\u00B1' to 0xF1,
+          '\u201C' to 0xF2,
+          '\u00BE' to 0xF3,
+          '\u00B6' to 0xF4,
+          '\u00A7' to 0xF5,
+          '\u00F7' to 0xF6,
+          '\u201E' to 0xF7,
+          '\u00B0' to 0xF8,
+          '\u2219' to 0xF9,
+          '\u00B7' to 0xFA,
+          '\u00B9' to 0xFB,
+          '\u00B3' to 0xFC,
+          '\u00B2' to 0xFD,
+          '\u25A0' to 0xFE,
+          '\u00A0' to 0xFF,
+      )
+}
